ENH:Function segmentation
The checkforwin function is subdivided into checkHorizontal,checkVertical,checkDiagonal,checkDraw.

The checkHorizontal function checks the completion of the horizontal line.
The checkVertical function checks the completion of the vertical line.
The checkDiagonal function checks the completion of the diagonal line.
The checkDraw function checks if all the fields are filled.